agent/requirements.txt
python-dotenv
coloredlogs
robyn==0.65.0
#addict
#SQLAlchemy
#requests
crawl4ai
openai
pydantic
rich
aioipfs
torrentp